requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pillow==10.3.0
python-multipart==0.0.9
websockets==12.0
torchvision
diffusers
transformers
accelerate
safetensors
open_clip_torch
open3d
trimesh
numpy
scipy
tqdm
orjson
pyyaml